% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbial-package.R
\docType{package}
\name{adverbial-package}
\alias{adverbial}
\alias{adverbial-package}
\title{adverbial: Enhanced Partialised and Composed Functions}
\description{
Provides new_partialised() and new_composed(), which extend the partial() and compose() functions of 'purrr' to make it easier to extract and replace arguments and functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/adverbial}
  \item Report bugs at \url{https://github.com/UchidaMizuki/adverbial/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
