% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialised.R
\name{arg}
\alias{arg}
\alias{arg<-}
\title{Arguments for partialised functions}
\usage{
arg(x, which)

arg(x, which) <- value
}
\arguments{
\item{x}{Partialised function.}

\item{which}{A non-empty character string specifying which argument is to be
accessed.}

\item{value}{An object, the new value of the argument.}
}
\value{
\code{arg} returns an argument.
}
\description{
Arguments for partialised functions
}
