% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialised.R
\name{arguments}
\alias{arguments}
\alias{arguments<-}
\title{Argument lists for partialised functions}
\usage{
arguments(x)

arguments(x) <- value
}
\arguments{
\item{x}{Partialised function.}

\item{value}{A list of arguments.}
}
\value{
\code{arguments()} returns a list of arguments.
}
\description{
Argument lists for partialised functions
}
