% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialised.R
\name{extract}
\alias{extract}
\alias{[.partialised}
\alias{[<-.partialised}
\alias{[[.partialised}
\alias{[[<-.partialised}
\alias{$.partialised}
\alias{$<-.partialised}
\title{Extract or replace arguments for partialised functions}
\usage{
\method{[}{partialised}(x, i, ...)

\method{[}{partialised}(x, i) <- value

\method{[[}{partialised}(x, i, ...)

\method{[[}{partialised}(x, i) <- value

\method{$}{partialised}(x, i)

\method{$}{partialised}(x, i) <- value
}
\arguments{
\item{x}{Partialised function.}

\item{i}{Indices specifying arguments to extract or replace.}

\item{...}{Additional arguments.}

\item{value}{An object, the new value of the argument.}
}
\value{
\code{[}, \code{[[} and \code{$} return arguments.
}
\description{
Extract or replace arguments for partialised functions
}
