% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed.R
\name{functions}
\alias{functions}
\alias{functions<-}
\title{Function lists for composed functions}
\usage{
functions(x)

functions(x) <- value
}
\arguments{
\item{x}{Composed function.}

\item{value}{A list of functions.}
}
\value{
\code{functions()} returns a list of functions.
}
\description{
Function lists for composed functions
}
