% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialised-package.R
\docType{package}
\name{partialised-package}
\alias{partialised}
\alias{partialised-package}
\title{partialised: Partialised Functions}
\description{
Provides a 'partialised' class that extends the partialising function of 'purrr' by making it easier to change the arguments. This is similar to the function-like object in 'Julia' (\url{https://docs.julialang.org/en/v1/manual/methods/#Function-like-objects}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/partialised}
  \item Report bugs at \url{https://github.com/UchidaMizuki/partialised/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
